feat: add files-zip function for zipping files with a search parameter
This commit adds a new function called files-zip to the file_tools.sh script. The files-zip function allows users to zip files in a directory based on a search parameter. It takes two arguments: the search parameter and the name of the zip file. If either of the arguments is missing, an error message is displayed.

The files-zip function uses the find command to search for files in the current directory that match the search parameter. It then pipes the output to the zip command to create a zip file with the specified name.